requirements.txt
appdirs==1.4.4
astroid==3.3.6
autopep8==2.3.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
dill==0.3.9
greenlet==3.1.1
idna==3.10
isort==5.13.2
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
pip==24.2
platformdirs==4.3.6
psycopg2==2.9.10
pycodestyle==2.12.1
pygubu==0.35.6
pygubu-designer==0.39.3
pylint==3.3.2
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
requests==2.32.3
screeninfo==0.8.1
setuptools==75.1.0
SQLAlchemy==2.0.36
telebot==0.0.5
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
